Create Estudia con Yosselin 1.1
This is the second part of my web creation and I want to show how I do my best while I'm studying the introduction at this signature. I know that my select of color it's not the best, but I know that I'm in progress. Finally I hope that in maybe a few months I will become in a great IT devoloper and 'll learn the all tips or hacks to do easier this files.