Create Estudia con Yosselin 1.2
In this case I wanted to implement a several more options to select. For example I create a house bottom to make easier the turn back If you are at some select like "Matemáticas". Then I put a site to make questions and answers If someone needs to ask me something or just give me a suggest.